separated out the base logic
so it's easier to port it in go
