--- src/constant/logger.ts
@@ -1,15 +1,0 @@
-export enum LogTypes {
-	ERROR = "error",
-	INFO = "info",
-	SUCCESS = "success",
-	PROGRESS = "progress",
-	DEBUG = "debug",
-	STEP = "step",
-}
-
-export enum LogProgressStatus {
-	QUEUED = "Queued",
-	WORKING = "Working",
-	DONE = "Done",
-	ERROR = "Error",
-}
