--- src/types/logger.ts
@@ -1,1 +1,0 @@
-export type { LogProgressStatus } from "@/constant/logger";
